feat: [MOM-51] Added type with settings and implemented logic for updating settings
