+++ src/context/DataContext.tsx
@@ -39,6 +39,8 @@ export const DataProvider = ({ children }: DataProviderProps) => {
                 responseData.Background = res.Background
                 responseData.Exchange = res.Exchange
                 responseData.Market = res.Market
+                responseData.Settings = res.Settings
+                localStorage.setItem("data", JSON.stringify(res));
                 setData(res)
               } else {
                 console.log("Something went wrong! Data is empty. Initial data will be displayed.")
