+++ src/models/response.types.ts
@@ -1,8 +1,16 @@
 import { Background } from "./Background/background.types"
-import { Exchange } from "./Exchange/exchange.types"
+import { Exchange, ExchangeRequest } from "./Exchange/exchange.types"
 import { Quote } from "./Quote/quote.types"
-import { Market } from "./Market/market.types"
-import { Weather } from "./Weather/weather.typse"
+import { Market, MarketRequest } from "./Market/market.types"
+import { Weather, WeatherRequest } from "./Weather/weather.typse"
+
+export type Settings = {
+    Weather: WeatherRequest
+    Quote: Quote
+    Background: Background
+    Exchange: ExchangeRequest
+    Market: MarketRequest
+}
 
 export type ResponseData = {
     Weather: Weather
@@ -10,4 +18,5 @@ export type ResponseData = {
     Background: Background
     Exchange: Exchange
     Market: Market
+    Settings: Settings
 }
