+++ src/pages/Home.page.tsx
@@ -1,19 +1,18 @@
 import { useContext, useState } from "react";
-import Button from "@mui/material/Button";
+import { Modal, Button, IconButton, Box } from "@mui/material";
 import FollowTheSignsIcon from '@mui/icons-material/FollowTheSigns';
-import IconButton from '@mui/material/IconButton';
 import SettingsIcon from '@mui/icons-material/Settings';
 
 import { UserContext } from "../context/UserContext";
 import { DataContext } from "../context/DataContext";
 import WeatherPage from "./Weather.page";
+import MarketPage from "./Market.page";
 import ExchangePage from "./Exchange.page";
 import QuoteCard from "../features/quote/components/quote-card.component";
 import Greeting from "../features/greeting/components/greeting.component";
+import SettingCard from "../features/settings/components/settings-card.component";
 
 import "./Home.style.css";
-import MarketPage from "./Market.page";
-import SettingCard from "../features/settings/components/settings-card.component";
 
 
 const HomePage = () => {
@@ -23,6 +22,7 @@ const HomePage = () => {
   const [isOpenSettings, setIsOpenSettings] = useState(false)
   const [is24HourFormat, setIs24HourFormat] = useState<boolean>(true);
 
+  const handleClose = () => setIsOpenSettings(false)
   
   const clickHandler = () => {
     localStorage.removeItem("token");
@@ -44,12 +44,19 @@ const HomePage = () => {
       <div className="background-text">
         <a href={data.Background.source_url} target="_blank">{data.Background.photographer} / {data.Background.source}</a>
       </div>
-      {isOpenSettings && <SettingCard  is24HourFormat={is24HourFormat} setIs24HourFormat={setIs24HourFormat} />}
-      <div className="setting">
+      <Modal
+        open={isOpenSettings}
+        onClose={handleClose}
+        aria-labelledby="modal-modal-title"
+        aria-describedby="modal-modal-description"
+      >
+        <SettingCard is24HourFormat={is24HourFormat} setIs24HourFormat={setIs24HourFormat} />
+      </Modal>
+      <Box className="setting">
         <IconButton  aria-label="skip" onClick={optionsHandler}>
           <SettingsIcon />
         </IconButton>
-      </div>
+      </Box>
     </div>
   )
 };
