+++ src/utils/initialData.ts
@@ -1,6 +1,10 @@
 import { ResponseData } from "../models/response.types";
 
-export const initialData: ResponseData = {
+const localData = localStorage.getItem("data");
+
+const data = JSON.parse(localData as string) as ResponseData
+
+export const initialData: ResponseData = data ? data : {
   Weather: {
       icon: "string",
       feels_like: 0,
@@ -34,5 +38,18 @@ export const initialData: ResponseData = {
     price: "string",
     change: "string",
     change_percent: "string"
+  },
+  Settings: {
+    Weather: {city: "string", source: "string"},
+    Quote: {author: "string", content: "string"},
+    Background: {
+      alt: "string",
+      image: "string",
+      photographer: "string",
+      source: "string",
+      source_url: "string",
+    },
+    Exchange: {from: "string", to: "string", source: "string"},
+    Market: {symbol: "string"}
   }
 }
